requirements.txt
beautifulsoup4
requests
openai
streamlit
python-dotenv
